+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Buffomat Classic 2023.11.4
+
+- Single rows for each food buff are now grouped per stat they provide, including all levels (helps while leveling). Highest 
+  available in your bag will be eaten first (and a setting to choose lowest first).
+  - New grouped buffs are marked with green text
+  - New option (last in the "Scan Options") to prefer either highest available 
+    item, or lowest. Lowest is useful when leveling, to consume your lowest level consumables first.
+  - Clicking the icon in the buff list to print all items, which can provide this buff, from lowest to highest.
+  - Added food eating auras to ignore list to hide the food buff task while eating, so it doesn't get eaten multiple times.
+- [Known Bug] Soon after resurrecting under the flask effect, Buffomat can show a task for reflasking. It disappears a second or two later.
+- In raid, staggered per-group updates now will skip empty 5man groups, which will result in faster updates overall.
+
 ## Buffomat Classic 2023.11.1
 
 - Hide consumables which have required level above player. If multiple items provide the buff, choose the lowest level